Added a DummyHeaderParser to facilitate CEF connectivity test.
The parser simply returns an empty "no.difi.oxalis.api.header.Header"
object. This is NOT recommended for production use. If you intend to
use Oxalis outside of the PEPPOL context is is HIGHLY recommended that
you familiarize yourself with the extension points of Oxalis and
implements the affected interfaces.
https://github.com/difi/oxalis/blob/master/doc/extension-points.adoc